Remove CommonMailer
